requirements.txt
spacy>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
nltk>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
kaggle>=1.5.0
python-dateutil>=2.8.0
regex>=2023.0.0
wordcloud>=1.9.0
plotly>=5.17.0

